LPD-23882 Move AMD stuff to amd dir (instead of npmscripts)
See https://github.com/liferay-frontend/liferay-portal/pull/4125#discussion_r1596281870
