+++ modules/_node-scripts/amd/writeAMD2ESMBridges.mjs
