+++ modules/_node-scripts/amd/writeManifestJson.mjs
