+++ modules/_node-scripts/amd/writePackageJson.mjs
